// generated/src/app/api-models/special-contract-detail.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class SpecialContractDetail {
  seqSpecCont: number;
  seqSpecContDtl: number;
  itemDescription: string;
  procedureCode: string;
  allowedAmt: number;
  allowedReason: string;
  multAmtByQty: string;
  pctOfBilled: number;
  pctBilledReason: string;
  useGreaterOrLess: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}